[irods/irods#7265] Minor reorganization + CMake TLC
- Project tree structure reorganized to resemble that of main irods repo.
- Added quotes around paths in CMake files.
- Removed CMAKE_SHARED_LINKER_FLAGS_INIT and
  CMAKE_SHARED_LINKER_FLAGS_RELEASE_INIT, as no shared libraries are built.
- Removed unused postinst script for nonexistent data_world package
- Removed rpm package dependency declaration on openssl
- Fixed postinst CPack variables
