+++ indexing/packaging/test_plugin_indexing.py
